+++ resolwe_bio/tests/processes/test_support_processors.py
@@ -19,26 +19,34 @@ class SupportProcessorTestCase(BioProcessTestCase):
 
     def test_bam_split(self):
         bam = self.prepare_bam('hybrid.bam')
-        header = self.run_process('upload-header-sam', {'src': 'mm10_header.sam'})
-        header2 = self.run_process('upload-header-sam', {'src': 'dm6_header.sam'})
 
         inputs = {
             'bam': bam.id,
-            'header': header.id,
-            'header2': header2.id,
             'organism': 'mm10',
             'organism2': 'dm6'
         }
         self.run_process('bam-split', inputs)
 
-        split_data = list(Data.objects.all())
-        bam1 = split_data[-2]
-        bam2 = split_data[-1]
+        bam2, bam1 = Data.objects.order_by('-id')[0:2]
 
         self.assertFile(bam1, 'bam', 'hybrid_mm10.bam')
         self.assertFile(bam1, 'bai', 'hybrid_mm10.bam.bai')
         self.assertFile(bam2, 'bam', 'hybrid_dm6.bam')
-        self.assertFile(bam2, 'bam', 'hybrid_dm6.bam.bai')
+        self.assertFile(bam2, 'bai', 'hybrid_dm6.bam.bai')
+
+        header = self.run_process('upload-header-sam', {'src': 'mm10_header.sam'})
+        header2 = self.run_process('upload-header-sam', {'src': 'dm6_header.sam'})
+
+        inputs['header'] = header.id
+        inputs['header2'] = header2.id
+        self.run_process('bam-split', inputs)
+
+        bam2, bam1 = Data.objects.order_by('-id')[0:2]
+
+        self.assertFile(bam1, 'bam', 'hybrid_mm10.bam')
+        self.assertFile(bam1, 'bai', 'hybrid_mm10.bam.bai')
+        self.assertFile(bam2, 'bam', 'hybrid_dm6.bam')
+        self.assertFile(bam2, 'bai', 'hybrid_dm6.bam.bai')
 
     def test_feature_location(self):
         inputs = {'src': 'mm10_small.gtf.gz', 'source': 'UCSC'}